DF-6043 cubemaps changes

+++ Tools/ResourceEditor/Classes/Qt/CubemapEditor/CubeMapTextureBrowser.cpp
@@ -313,18 +313,21 @@ void CubeMapTextureBrowser::OnDeleteSelectedItemsClicked()
 				FilePath fp = item->data(CUBELIST_DELEGATE_ITEMFULLPATH).toString().toStdString();
 				if(fp.Exists())
 				{
-					DAVA::Vector<DAVA::String> faceNames;
+					DAVA::Vector<DAVA::FilePath> faceNames;
 					CubemapUtils::GenerateFaceNames(fp.GetAbsolutePathname(), faceNames);
 					for(size_t faceIndex = 0; faceIndex < faceNames.size(); ++faceIndex)
 					{
+                        if (faceNames[faceIndex].IsEmpty())
+                            continue;
+
 						FilePath hackTex = faceNames[faceIndex];
 						hackTex.ReplaceExtension(".tex");
 						
 						QFile::remove(hackTex.GetAbsolutePathname().c_str());
-						bool removeResult = QFile::remove(faceNames[faceIndex].c_str());
+						bool removeResult = QFile::remove(faceNames[faceIndex].GetAbsolutePathname().c_str());
 						if(!removeResult)
 						{
-							failedToRemove.push_back(faceNames[faceIndex]);
+                            failedToRemove.push_back(faceNames[faceIndex].GetAbsolutePathname().c_str());
 						}
 					}
 					
@@ -367,12 +370,15 @@ bool CubeMapTextureBrowser::ValidateTextureAndFillThumbnails(DAVA::FilePath& fp,
 	
 	int width = 0;
 	int height = 0;
-	DAVA::Vector<DAVA::String> faceNames;
+	DAVA::Vector<DAVA::FilePath> faceNames;
 	CubemapUtils::GenerateFaceNames(fp.GetAbsolutePathname(), faceNames);
 	for(size_t i = 0; i < faceNames.size(); ++i)
 	{
+        if (faceNames[i].IsEmpty())
+            continue;
+
 		QImage faceImage;
-		if(!faceImage.load(faceNames[i].c_str())) //file must be present
+		if(!faceImage.load(faceNames[i].GetAbsolutePathname().c_str())) //file must be present
 		{
 			result = false;
 		}
